rename releng's hubris module to hubris_legacy
